Refactored ResetPassword feature into separate component, adjusted route @ App.js, add confirmation alert
