+++ src/components/ResetPassword/EnterNewPassword.js
@@ -1,16 +1,39 @@
 /** @jsx jsx */
-import { css, jsx } from "@emotion/core";
+import { jsx } from "@emotion/core";
 import styled from "@emotion/styled";
 import { useState } from "react";
-import { Redirect } from "react-router-dom";
-import { request } from "https";
+import { withRouter } from "react-router-dom";
+import SimpleReactValidator from "simple-react-validator";
 
-export default function ResetPassword(props) {
+function EnterNewPassword(props) {
   const [resetState, setLoginState] = useState({});
+
+  const validator = new SimpleReactValidator({
+    validators: {
+      password: {
+        message:
+          "Password must include at least one capital letter and one number",
+        rule: (val, params, validator) => {
+          return (
+            validator.helpers.testRegex(
+              val,
+              /^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])/
+            ) && params.indexOf(val) === -1
+          );
+        },
+        messageReplace: (message, params) =>
+          message.replace(":values", validator.helpers.toSentence(params)), // optional
+        required: true, // optional
+      },
+    },
+  });
 
   const handleSubmit = (e) => {
     e.preventDefault();
-    onSubmit(resetState);
+    if (validator.fieldValid("password")) {
+      // console.log("Valid pass", validator.fieldValid("password"));
+      onSubmit(resetState);
+    }
   };
 
   const handleChange = (event) => {
@@ -20,12 +43,13 @@ export default function ResetPassword(props) {
       [name]: value,
     });
   };
-  console.log("props.history.", props.history);
 
   const onSubmit = (data) => {
-    const { email } = data;
-    props.resetPassword(email);
-    props.history.replace(`/Investors/login`);
+    const { password } = data;
+    const token = props.location.pathname.split("/")[2];
+    props.updatePassword(token, password);
+    alert("Password changed, please login.");
+    props.history.replace(`/`);
   };
 
   if (props)
@@ -33,21 +57,29 @@ export default function ResetPassword(props) {
       <Container>
         <LeftSide>
           <div>
-            <h3>Reset Password</h3>
+            <h3>Enter New Password</h3>
           </div>
         </LeftSide>
-
         <RightSide>
           <form onSubmit={handleSubmit}>
-            <label>Email</label>
+            <label>New password</label>
             <input
-              type="email"
-              name="email"
-              value={resetState.email || ""}
+              type="password"
+              required="required"
+              name="password"
+              value={resetState.password || ""}
               onChange={handleChange}
             />
-            <br />
-            <button type="submit">Reset Password</button>
+            {resetState.password &&
+              validator.message(
+                "password",
+                resetState.password,
+                `required|password:|min:8`
+              )}
+            <ErrorMsg>
+              <p>{validator.errorMessages.password}</p>
+            </ErrorMsg>
+            <button type="submit">Submit</button>
           </form>
         </RightSide>
       </Container>
@@ -55,13 +87,13 @@ export default function ResetPassword(props) {
   else return <div></div>;
 }
 
-const Logo = styled.img`
-  position: absolute;
-  left: 0;
-  right: 0;
-  top: 50%;
-  margin: -300px auto auto auto;
-  height: 70px;
+export default withRouter(EnterNewPassword);
+
+const ErrorMsg = styled.div`
+  color: #ff0000;
+  font-size: 0.7em;
+  padding-left: 2.7em;
+  padding-right: 1em;
 `;
 
 const LeftSide = styled.div`
@@ -161,8 +193,8 @@ const RightSide = styled.div`
     position: relative;
     float: right;
     right: 10%;
-    width: 50%;
-    height: 50px;
+    width: 30%;
+    height: 30px;
     line-height: 30px;
     font-size: 12px;
     color: #233949;
